# RecipeBookWebClient/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "RecipeBookWebApi",
    "version": "1.0"
  },
  "paths": {
    "/api/Dishes": {
      "get": {
        "tags": [
          "Dishes"
        ],
        "parameters": [
          {
            "name": "Category",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DishCategory"
            }
          },
          {
            "name": "Flags",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/DishFlags"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DishResponseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DishResponseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DishResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Dishes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DishRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DishRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Dishes/{id}": {
      "get": {
        "tags": [
          "Dishes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DishResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Dishes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DishRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DishRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DishRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Dishes"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Dishes/calculate-nutrition": {
      "post": {
        "tags": [
          "Dishes"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NutritionCalculateRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/NutritionCalculateRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/NutritionCalculateRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/NutritionResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Products": {
      "get": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "Category",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProductCategory"
            }
          },
          {
            "name": "CookingRequired",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/CookingRequired"
            }
          },
          {
            "name": "Flags",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProductFlags"
            }
          },
          {
            "name": "Search",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "SortBy",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ProductSortParam"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponseDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponseDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponseDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Products"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/Products/{id}": {
      "get": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequestDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequestDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Products"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "CookingRequired": {
        "enum": [
          "ГотовыйКУпотреблению",
          "Полуфабрикат",
          "ТребуетПриготовления"
        ],
        "type": "string"
      },
      "DishCategory": {
        "enum": [
          "Десерт",
          "Первое",
          "Второе",
          "Напиток",
          "Салат",
          "Суп",
          "Перекус"
        ],
        "type": "string"
      },
      "DishFlags": {
        "enum": [
          "None",
          "Веган",
          "БезГлютена",
          "БезСахара"
        ],
        "type": "string"
      },
      "DishIngredientDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "integer",
            "format": "int32"
          },
          "amount": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "DishRequestDto": {
        "required": [
          "calories",
          "carbs",
          "fats",
          "ingredients",
          "name",
          "portionSize",
          "proteins"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 2,
            "type": "string"
          },
          "photos": {
            "maxItems": 5,
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "calories": {
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "proteins": {
            "maximum": 100,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "fats": {
            "maximum": 100,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "carbs": {
            "maximum": 100,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "ingredients": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DishIngredientDto"
            }
          },
          "portionSize": {
            "minimum": 0.1,
            "type": "number",
            "format": "double"
          },
          "category": {
            "$ref": "#/components/schemas/DishCategory"
          },
          "flags": {
            "$ref": "#/components/schemas/DishFlags"
          }
        },
        "additionalProperties": false
      },
      "DishResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "photos": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "calories": {
            "type": "number",
            "format": "double"
          },
          "proteins": {
            "type": "number",
            "format": "double"
          },
          "fats": {
            "type": "number",
            "format": "double"
          },
          "carbs": {
            "type": "number",
            "format": "double"
          },
          "ingredients": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DishIngredientDto"
            },
            "nullable": true
          },
          "portionSize": {
            "type": "number",
            "format": "double"
          },
          "category": {
            "$ref": "#/components/schemas/DishCategory"
          },
          "flags": {
            "$ref": "#/components/schemas/DishFlags"
          }
        },
        "additionalProperties": false
      },
      "NutritionCalculateRequestDto": {
        "required": [
          "ingredients"
        ],
        "type": "object",
        "properties": {
          "ingredients": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DishIngredientDto"
            }
          }
        },
        "additionalProperties": false
      },
      "NutritionResponseDto": {
        "type": "object",
        "properties": {
          "calories": {
            "type": "number",
            "format": "double"
          },
          "proteins": {
            "type": "number",
            "format": "double"
          },
          "fats": {
            "type": "number",
            "format": "double"
          },
          "carbs": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "ProductCategory": {
        "enum": [
          "Замороженный",
          "Мясной",
          "Овощи",
          "Зелень",
          "Специи",
          "Крупы",
          "Консервы",
          "Жидкость",
          "Сладости"
        ],
        "type": "string"
      },
      "ProductFlags": {
        "enum": [
          "None",
          "Веган",
          "БезГлютена",
          "БезСахара"
        ],
        "type": "string"
      },
      "ProductRequestDto": {
        "required": [
          "calories",
          "carbs",
          "category",
          "cookingRequired",
          "fats",
          "name",
          "proteins"
        ],
        "type": "object",
        "properties": {
          "name": {
            "minLength": 2,
            "type": "string"
          },
          "photos": {
            "maxItems": 5,
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "calories": {
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "proteins": {
            "maximum": 100,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "fats": {
            "maximum": 100,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "carbs": {
            "maximum": 100,
            "minimum": 0,
            "type": "number",
            "format": "double"
          },
          "composition": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/ProductCategory"
          },
          "cookingRequired": {
            "$ref": "#/components/schemas/CookingRequired"
          },
          "flags": {
            "$ref": "#/components/schemas/ProductFlags"
          }
        },
        "additionalProperties": false
      },
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "photos": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "calories": {
            "type": "number",
            "format": "double"
          },
          "proteins": {
            "type": "number",
            "format": "double"
          },
          "fats": {
            "type": "number",
            "format": "double"
          },
          "carbs": {
            "type": "number",
            "format": "double"
          },
          "composition": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "$ref": "#/components/schemas/ProductCategory"
          },
          "cookingRequired": {
            "$ref": "#/components/schemas/CookingRequired"
          },
          "flags": {
            "$ref": "#/components/schemas/ProductFlags"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "ProductSortParam": {
        "enum": [
          "Name",
          "Calories",
          "Proteins",
          "Fats",
          "Carbs"
        ],
        "type": "string"
      }
    }
  }
}